chore: add netboot docker server
It will be a lot easier to do development on the verdin without having
to constantly bounce around between recovery mode and flashing the
internal mmc. This adds a docker container and instructions for host
system standup (it is not that isolated) that can connect directly to a
verdin via an ethernet interface, and provide a networked bootloader
chain given a u-boot config on the verdin that accepts it.

The docker container, when bound to the appropriate iface, provides a
dhcp server, which also provides a bootp kernel download; and then an
nfs filesystem server for the root filesystem.

When this all works, with the makefile provided, it should allow devs to
carve out a directory for a filesystem unzipped from an openembedded
build, and boot a verdin off their local machine, with the ability to
much more easily change files "on" the device.