build: use github actions self-hosted runners to build (#30)
Our AWS codebuild setup works, but double-bills us and is annoying to see logs from and modify. Plus, we can't really change the machines that the build runs on.

Instead, use new infrastructure that builds on runners provided by EC2 autoscaling groups and thus free to be any kind of EC2 instance; and that pushes builds to s3 buckets that actually get nice URLs.