layers/meta-opentrons: new flags required for node
Kirkstone uses openssl v3, which node 16 doesn't actually have great
compatibility with. Our old recipe didn't have this issue I think
because dunfell was using openssl v2. To fix this, we need to get node
using the old openssl, which you would normally do through
a (patched-back-in) command line flag, but that's a flag to _node_, and
we're actually running yarn, so we have to pass it in through env vars,
and then it works.