meta-opentrons: add custom logo
This sets up the configuration for the kernel to display a splash
screen (if given the proper commandline). The image has to be in a magic
format (224 color ascii ppm) in a magic size (the largest multiple of 16
pixels that is smaller than the target image size) in a magic place and
now it is.