meta-opentrons: add u-boot append for env args
U-boot sets default environment values by compiling them into the
binary. It's good to keep these extremely minimal, and they do but also
they're still too much for what we want to do. For a splash screen, we
need to disable the video console and change some basic args, and the
compiled in settings prevent that. We could alter these in the boot
script but then we'd need some further signal beyond "is this variable
set or not" to determine whether the value was default, and this is simpler.