meta-opentrons: add userspace loading seq
When the system gets to userspace, it still takes a while to get the
robot app booted. We want to fill that time with some loading animation,
which will also be an animation that shows when the app crashes.

To do this, we can make the "animation" actually just an mp4. We can
install a utility called gstd, which is a daemon that accepts bus
commands from a frontend to build and control gstreamer pipelines, and
then we can make a service that uses it to loop the video without losing
frames - that's our animation.

So in this commit, we
- add an RDEP on gstd in the app launcher
- add a systemd service for gstd with appropriate env vars to talk to
weston
- add a systemd service for a loading screen that runs at boot and if
the app crashes (using explicit enable/disables in units rather than
Conflicts because otherwise it won't work correctly on boot)
- add a nice temporary video to take the place of the eventual loading
screen asset