opentrons-robot-server: store api package version
We were not storing the version of the api package that the robot server
relies on. This is used by the update server to respond with the version
of the loaded package, and without it we don't get a version there. On
the OT-2, this is provided by the separate api package recipe, but we've
folded that in to the robot server here so we need to do it in the same
recipe.

Closes RQA-577