layers/meta-opentrons: save logs on disk (#81)
We previously had storage on auto but didn't have a /var/logs/journal so
logs were going only to RAM with a limit of 64 megs. Now we'll make the
directory and store 4G of logs.

Annoyingly, even though we can change the directory and settings, we still won't
saving to disk because of things like /var/log being symlinked to
/var/volatile/log. So we make a unit that operates only if that symlink
exists and removes it and makes the actual journal directory. Seems to
work fine!

Co-authored-by: vegano1 <brayan.almonte@gmail.com>