layers/meta-opentrons: update loading screens
The new userspace loading video does not loop, which means it doesn't
glitch when transitioning from video to app. The new kernel loading
screen is darker which means we get a smooth color gradient from off to
running, and the new video transitions from the dark mode video to running.