opentrons-ssh-keys: always install file (#107)
We weren't installing the key in the release execution branch but were
still providing it in the OE recipe. At some point we should make the
build mode and project actual OE build variables that you can establish
recipe variants on but for now just put an empty file in there in the
other branch.