Update openembedded to kirkstone/tdx 6.3.0 (#110)
Updates the system to toradex upstream 6.3.0, including
- update submodules to toradex 6.3.0
- update jupyter to 2022.2
- update python to 3.10
- update setuptools to 62.3.1

Add the recipe from openembedded-core upstream (version unclear, maybe
mickledore) for setuptools 62.3.1, since there are some python packages
- specifically kiwisolver from meta-openembedded and argon2-cffi from
meta-jupyter - that require more recent setuptools to generate proper
dist-infos. Without it, with just the version in openembedded-core, they
generate dist-infos that are UNKNOWN-0.0.0 and conflict with each other
during image install. With 62.3.1, they generate kiwisolver-0.0.0 and
argon2-cffi-0.0.0, which is still annoying but okay for now.

- actually turn on the gpu

Why was this not in the examples and why did we think it was fine to not
have it

---------

Co-authored-by: Seth Foster <seth@opentrons.com>