move unit, fix options, remove packagegroup
This hopefully should be the last step to getting everything working and
sane. We don't need the packagegroup (something was breaking everything
anyway) and we need to fix up a couple things in the install process so
the unit goes in the same place as everything else, actually runs
jupyter directly so journald logging tags it properly, and we get the
user package path.