fix: correct devtools and fix app scripts
We tried to get nice readable scripts and environments in the app
runscript and systemd unit and it didn't work and it's because of the
sed thing and that's stupid so just write these values down in the files
and check them in.

Also fix the devtools allow-origin to be pluralized so it actually works.