chore: add an autoupdate script (#205)
The sources that we specify for the layers in openembedded are all taken
from a google repotool manifest run by toradex. To update our
openembedded, we have to update to the refs specified for all the layers
by the repotool manifest. This is an annoying and manual pain.

Instead, here's a script that will just update all the submodules to a
target ref of the repotool manifest. Run without --update to predict and
with --update to execute. You'll have to commit the submodule changes if
you want to afterwards.

This could theoretically be integrated into ci with some changes or a
check mode that just prints if it would change anything pretty easily.