refactor(meta-opentrons):  Add udev rule for system camera (#214)
Closes EXEC-1643

The USB camera that ships with the Flex seemingly gets assigned /dev/video2 as its device node, but we'd like to rename the node to something more semantic, like ot_system_camera. Because this camera is not a module but is a device supported by the Flex, we'll need some sort of new naming convention here, and ot_system is the proposal.