oe-core: move data out of build
There's lots of annoying thigns about openembedded but this annoys me
perhaps the most, so let's just stop doing it.

Instead of checking the configuration files into build, just put them in
a conf/ directory and copy them in just in time when we start the docker
container. This means we can gitignore all of build and just rm -rf it
whenever we want to restart the build. Pretty convenient!