meta-opentrons: revert bad kernel commit
Per
https://community.toradex.com/t/verdin-mini-m4-uart4-stops-with-kernel-boot/27112/3
kernel commit
https://git.toradex.com/linux-toradex.git/commit/?h=toradex_6.6-2.2.x-imx&id=4f6c9832613b421e56bad2f9e4e2767e43e1f97c
causes the pmic to down sometimes; in the case of the question, when you
upped the embedded m4 core but in our case just during boot all the
time. This turns off half the devices on the SOM including things like
usb. Reverting the commit per the question (which is a little more
involved than git revert because in between that commit and now that
entire kernel driver subtree changed names) fixes the issue.