layers/meta-opentrons: various gstreamer fixes
We need to explicitly install our plugins now for whatever reason; also
gstd as a daemon works differently and needs to know the wayland display
or it segfaults when you start a pipeline, which is very easy to figure
out. Also we should sigkill it because it ignores sigterm when it's playing.