meta-opentrons: remove sentry cli before pack (#218)
If you remove the sentry cli binaries (that are incompatible with our
arch) in fakeroot from the install dir, then you actually invalidate the
asar, which has the entries for everything in app.asar.unpacked as well
as the stuff in app.asar in its registry so it can present them
seamlessly. Instead, remove them from the "source" node_modules before
we pack the app so they just never get packed in the first place.